Add a warning to the 5.7.4 release notes about a bug in LogixNG
Type conversion to Boolean uses Jython rules. JMRI 5.7.4 has new features there this doesn't work. So in JMRI 5.7.5, this will be changed so that the string "true" (case insensitive) is converted to true and the string "false" (case insensitive) is converted to false and that other string values including null throws an exception.